API/weatherapi/requirements.txt
asn1crypto==1.5.1
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
idna==3.6
numpy==1.24.4
pandas==2.0.3
pg8000==1.30.4
pipreqs==0.4.13
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scramp==1.4.4
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
yarg==0.1.9

